requirements.txt
fastapi
uvicorn
pydantic
pyodbc
numpy
pandas
scikit-learn
python-dotenv
jinja2
python-multipart
